+++ examples/get_notes.rb
@@ -2,7 +2,7 @@ require_relative "../lib/xolphin/api"
 require_relative "_credentials"
 
 client = Xolphin::Api::Client.new(Credentials.username, Credentials.password, true)
-result = client.request.getNotes(960003015)
+result = client.request.getNotes(960_003_015)
 puts result.inspect
 result.notes.each do |note|
   puts note.message_body, note.created_at