+++ examples/send_csa_email.rb
@@ -1,5 +1,5 @@
 require_relative "../lib/xolphin/api"
 
 client = Xolphin::Api::Client.new(Credentials.username, Credentials.password, true)
-result = client.request.sendComodoSAEmail(960003013, 'nikita@xolphin.nl');
+result = client.request.sendComodoSAEmail(960_003_013, 'nikita@xolphin.nl')
 puts result.inspect