+++ lib/xolphin/api/endpoint/request.rb
@@ -44,8 +44,8 @@ module Xolphin
           Xolphin::Api::Responses::Request.new(result)
         end
 
-        def createEE()
-          Xolphin::Api::CertificateRequests::CreateEERequest.new()
+        def createEE
+          Xolphin::Api::CertificateRequests::CreateEERequest.new
         end
 
         def sendEE(request)
@@ -55,30 +55,34 @@ module Xolphin
         end
 
         def sendNote(id, note)
-          @http.post("/requests/#{id}/notes",{'message' => note})
+          @http.post("/requests/#{id}/notes", 'message' => note)
         end
 
         def getNotes(id)
           result = @http.get("/requests/#{id}/notes")
-          
+
           Xolphin::Api::Responses::Notes.new(result)
         end
 
         def sendComodoSAEmail(id, to, language = nil)
-          @http.post("requests/#{id}/sa",{'sa_email' => to, 'language' => language})
+          @http.post("requests/#{id}/sa", 'sa_email' => to, 'language' => language)
         end
 
         def upload(id, document, description = nil)
-          @http.post("requests/#{id}/upload-document",{'document' => document, 'description' => description})
+          @http.post("requests/#{id}/upload-document", 'document' => document, 'description' => description)
         end
 
         def scheduleValidationCall(id, date_time)
           t = Time.parse(date_time)
-          @http.post("requests/#{id}/schedule-validation-call",{'date' => t.strftime("%Y-%m-%d"), 'time' => t.strftime("%H:%M")})
+          @http.post("requests/#{id}/schedule-validation-call", 'date' => t.strftime("%Y-%m-%d"), 'time' => t.strftime("%H:%M"))
         end
 
         def retryDCV(id, domain, dcv_type, email = nil)
-          @http.post("requests/#{id}/retry-dcv",{"domain" => domain, "dcvType" => dcv_type, "email" => email})
+          @http.post("requests/#{id}/retry-dcv", "domain" => domain, "dcvType" => dcv_type, "email" => email)
+        end
+
+        def cancel(id, reason = "")
+          @http.post("requests/#{id}/cancel", "reason" => reason)
         end
       end
     end