+++ lib/xolphin/api/invoice_format_type.rb
@@ -1,0 +1,8 @@
+module Xolphin
+  module Api
+    class InvoiceFormatType
+      PDF = "PDF"
+      UBL = "UBL"
+    end
+  end
+end