+++ lib/xolphin/api/responses/note.rb
@@ -8,6 +8,10 @@ module Xolphin
           super(data)
         end
 
+        def id
+          @data["id"]
+        end
+
         def contact
           @data["contact"]
         end
@@ -30,6 +34,10 @@ module Xolphin
 
         def created_at
           Time.parse(@data["createdAt"])
+        end
+
+        def end_customer
+          @data['endCustomer']
         end
       end
     end