+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [1.8.0] - 2020-03-02
+### Added
+- Certificate Requests: Support the new validation information from Comodo
+- Certificate Requests: Added method for cancelling a request
+- Certificate Requests: Added some missing properties
+- Certificate Requests: Added support for scheduling validation calls
+- Invoices: Added invoices endpoints
+
 ## [1.5.0] - 2017-06-07
 ### Added
 - Certificates: Reissue and renewal of certificates